Enable network IPMI for all network interfaces
The network IPMI is configured on a per channel basis to meet the requirements
of IPMI specification. The netipmid instance binds to a interface (eg eth0).
When a VLAN interface is created, it is a new interface and there is no netipmid
instance for the VLAN interface.

To workaround this restriction, this patch makes the changes to make the eth0
instance of netipmid listen on *.623. This will cater for the VLAN interfaces.
Since there is only one network interface on Witherspoon. The channel settings
of eth0 are inherited by the VLAN interfaces.

This fix doesn't qualify for upstreaming and it needs to address for cases where
there are multiple interfaces.

Signed-off-by: Tom Joseph <tomjoseph@in.ibm.com>