Require a valid session state prior to executing command
This is a manual edit of commit cfb34ca17cbe54514f3a77cf37afea68d2a875cc
into the OP940 branch.  git cherry-pick gave bad results.
The original commit message is below:

The execution of commands should work fine for pre-session commands
and for session commands that have the session in a valid state
(not inactive or tearDownInProgress). This prevents a session from
getting used after the close session command.

Tested: send a command after the session has been closed or re-use an
        old session ID. The BMC should ignore the request.

Change-Id: I112bbc3404ffcf90ab5358d2309672473662647a
Signed-off-by: Vernon Mauery <vernon.mauery@linux.intel.com>
Signed-off-by: Joseph Reynolds <joseph-reynolds@charter.net>