phosphor-ipmi-net: remove sys-subsystem-net-devices-%i.device dep
Commit 34b0b9c8 introduced the concept of waiting for the required
network device to be available before executing this service.

This works well except for systems where the ethernet devices are not
known at build time. For example, the p10bmc openbmc image supports
multiple machines, some with 2 ethernet ports, and others with just 1.

On the system with just 1 ethernet port, we end up hanging the boot to
BMC Ready by the 90s timeout while waiting for a non-existent eth1 port.

Unfortunately systemd does not look at the "ConditionPathExists" until
after it processes the "Requires" statements (even if you move it to the
first line in the service file).

I'm unable to recreate the issue seen in 34b0b9c8 on either of our
systems (1 with eth0 and eth1 and 1 with just eth0) so I'm not positive
this commit addresses the original concern, but it does at least get our
1 eth port system working again.

Tested:
- Confirmed eth0 netipmid comes up and eth1 does not due to Condition
  check.
- Confirmed that good path still works as expected on 2 eth port system

Change-Id: Ic224167db3b04871e1fb19def0f0d9f9f2b52e13
Signed-off-by: Andrew Geissler <geissonator@yahoo.com>